Add part 2, adapt part 1 code to new spec
Part 2 changed several aspects of the problem, but also shares a lot of the functionality.
I've reworked the approach to part 1 so both cases can be handled by the same code.
This mostly means a different way of defining the keypad arrays, handling valid array indices that don't represent an actual key, and allowing functions to handle different keypad layouts